Refacor Book Model
Refactor Book model from MongoDB to MySQL engine. Change annotations for entity. In facade refactor arguments for methods. Add Either to return errors on controller with correct status. Refactor unit tests for api.

+++ src/main/java/com/haredev/library/book/domain/BookFacade.java
@@ -1,8 +1,10 @@
 package com.haredev.library.book.domain;
 
 import com.haredev.library.book.controller.validation.BookValidation;
+import com.haredev.library.book.domain.api.BookError;
 import com.haredev.library.book.dto.BookCreateDto;
 import io.vavr.collection.Seq;
+import io.vavr.control.Either;
 import io.vavr.control.Option;
 import io.vavr.control.Validation;
 import lombok.AllArgsConstructor;
@@ -25,14 +27,17 @@ public class BookFacade {
         return bookValidation.validate(request);
     }
 
-    public Option<BookCreateDto> findBook(String bookId) {
+    public Option<BookCreateDto> findBook(Long bookId) {
         return bookRepository.findById(bookId).map(Book::response);
     }
 
     public List<BookCreateDto> fetchAllBooks() {
         return bookRepository.findAll().stream().map(Book::response).collect(Collectors.toList());
     }
-    public void removeBook(String bookId) {
-        bookRepository.deleteById(bookId);
+    public Either<BookError, BookError> removeBook(Long bookId) {
+        if(bookRepository.existsById(bookId)) {
+            bookRepository.deleteById(bookId);
+            return Either.right(BookError.OPERATION_SUCCESS); }
+        else return Either.left(BookError.BOOK_NOT_FOUND);
     }
 }
